feat(Backend): terminando de implementar el controller
corrigiendo errores de las entitys y haciendo las pruebas en postman

+++ FinTrack_API/src/main/java/com/fintrack/api/service/Impl/TransactionServiceImpl.java
@@ -2,6 +2,7 @@ package com.fintrack.api.service.Impl;
 
 import com.fintrack.api.exception.CategoryNotFoundException;
 import com.fintrack.api.exception.TransactionNotFountException;
+import com.fintrack.api.exception.UserNotFoundException;
 import com.fintrack.api.mapper.TransactionMapper;
 import com.fintrack.api.persistence.dto.request.TransactionRequest;
 import com.fintrack.api.persistence.dto.response.GenericResponse;
@@ -9,11 +10,14 @@ import com.fintrack.api.persistence.dto.response.PaginationResponse;
 import com.fintrack.api.persistence.dto.response.TransactionResponse;
 import com.fintrack.api.persistence.model.Category;
 import com.fintrack.api.persistence.model.Transaction;
+import com.fintrack.api.persistence.model.User;
 import com.fintrack.api.persistence.repository.CategoryRepository;
 import com.fintrack.api.persistence.repository.TransactionRepository;
+import com.fintrack.api.persistence.repository.UserRepository;
 import com.fintrack.api.service.TransactionService;
 import jakarta.transaction.Transactional;
 import java.util.List;
+import java.util.Objects;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -25,6 +29,7 @@ public class TransactionServiceImpl implements TransactionService {
 
   private final TransactionRepository transactionRepository;
   private final CategoryRepository categoryRepository;
+  private final UserRepository userRepository;
 
   @Override
   public PaginationResponse<TransactionResponse> findAll(Pageable pageable) {
@@ -55,6 +60,18 @@ public class TransactionServiceImpl implements TransactionService {
   }
 
   @Override
+  public List<TransactionResponse> findAllByCategoryName(String categoryName) {
+    List<Transaction> responses = transactionRepository.findAllByCategoryName(categoryName);
+    return TransactionMapper.toTransactionResponseList(responses);
+  }
+
+  @Override
+  public List<TransactionResponse> findAllByUserId(Long userId) {
+    List<Transaction> responses = transactionRepository.findAllByUserId(userId);
+    return TransactionMapper.toTransactionResponseList(responses);
+  }
+
+  @Override
   public TransactionResponse findById(Long id) {
     Transaction transaction = this.getTransaction(id);
     return TransactionMapper.toTransactionResponse(transaction);
@@ -72,19 +89,32 @@ public class TransactionServiceImpl implements TransactionService {
 
   @Override
   @Transactional
-  public TransactionResponse create(TransactionRequest request) {
+  public TransactionResponse create(TransactionRequest request, Long userId) {
     Category category = this.getCategory(request.categoryId());
     Transaction transaction = TransactionMapper.toTransaction(request);
+    User user = this.getUser(userId);
     transaction.setCategory(category);
+    transaction.setUser(user);
     Transaction savedTransaction = transactionRepository.save(transaction);
     return TransactionMapper.toTransactionResponse(savedTransaction);
   }
 
+  private User getUser(Long userId) {
+    return userRepository.findById(userId)
+        .orElseThrow(() -> new UserNotFoundException("User not found"));
+  }
+
   @Override
   @Transactional
-  public TransactionResponse update(Long id, TransactionRequest request) {
+  public TransactionResponse update(Long id, TransactionRequest request, Long userId) {
     Transaction transaction = this.getTransaction(id);
+    User user = this.getUser(userId);
     Category category = this.getCategory(request.categoryId());
+
+    if (!transaction.getUser().getId().equals(user.getId())) {
+      throw new UserNotFoundException("User not found");
+    }
+
     TransactionMapper.updateTransaction(transaction, request);
     transaction.setCategory(category);
     Transaction updatedTransaction = transactionRepository.save(transaction);
@@ -93,9 +123,16 @@ public class TransactionServiceImpl implements TransactionService {
 
   @Override
   @Transactional
-  public GenericResponse delete(Long id) {
+  public GenericResponse delete(Long id, Long userId) {
     Transaction transaction = this.getTransaction(id);
+    User user = this.getUser(userId);
+
+    if (!Objects.equals(transaction.getUser().getId(), user.getId())) {
+      throw new UserNotFoundException("User not found");
+    }
+
     transactionRepository.delete(transaction);
-    return new GenericResponse("Transaction eliminada correctamente");
+
+    return new GenericResponse("Transaction eliminada correctamente " + id);
   }
 }
